feat: complete NestJS implementation with PostgreSQL support - Implement all 13 core modules (User, Question, Answer, Comment, Vote, Tag, etc.) - Add complete RBAC permission system with RolePowerRel entity

+++ src/entities/user-role-rel.entity.ts
@@ -13,7 +13,7 @@ export class UserRoleRel {
   @UpdateDateColumn({ name: 'updated_at', comment: 'update time' })
   updatedAt: Date;
 
-  @Column({ name: 'user_id', type: 'bigint', default: 0, comment: 'user id' })
+  @Column({ name: 'user_id', type: 'int', default: 0, comment: 'user id' })
   userId: string;
 
   @Column({ name: 'role_id', type: 'int', default: 0, comment: 'role id' })
